fix(core): break core → portal type dependency to unblock agentbox build
The AgentBox Dockerfile (`Dockerfile.agentbox`) intentionally does not
copy `src/portal/` into the builder stage — it's a deliberate layering
rule (AgentBox runtime ↛ Portal UI/REST). But a recent PR added
type-only imports from `src/portal/cli-snapshot-api.ts` into
`src/core/`:

    src/core/agent-factory.ts:87,89            (import("../portal/..."))
    src/core/extensions/agent.ts:13            (import type from ../../portal/...)
    src/core/extensions/ls.ts:19               (import type from ../../portal/...)

GitHub Actions' `CI / Type Check` runs `tsc` on the full source tree
so it sees `src/portal/` and reports no error. But the siflow image
build runs `npx tsc -p tsconfig.agentbox.json` inside the Dockerfile
builder stage, which only has the AgentBox-scoped COPY subset — there
`src/portal/` does not exist, so tsc fails with:

    TS2307: Cannot find module '../portal/cli-snapshot-api.js'

All four imports are strictly type-only and only reference
`CliSnapshotAgentMeta` and `CliSnapshotActiveAgent`. Both interfaces
are plain POJOs with zero runtime dependencies.

Fix: relocate the two type contracts to `src/shared/cli-snapshot-types.ts`
(already copied by `Dockerfile.agentbox` and included by
`tsconfig.agentbox.json`), re-export them from the original location
in `src/portal/cli-snapshot-api.ts` so external consumers
(portal-web, siclaw-tui, etc.) don't need to change their imports,
and point the three `core/` files at the new location.

Zero runtime behavior change; restores AgentBox Docker build.

Verification:
- `npx tsc -p tsconfig.agentbox.json --noEmit` passes
- grep confirms no remaining `../portal/` imports under the AgentBox
  compile scope (core/, agentbox/, cron/, memory/, shared/, tools/,
  gateway/agentbox/, gateway/security/, and the handful of gateway
  top-level files listed in the Dockerfile).
